backend/requirements.txt
flask
flask-cors
python-dotenv
python-docx
pdfkit